tests/wpt/tests/custom-elements/registries/scoped-registry-append.html

<!DOCTYPE html>
<html>
<head>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>

<template id="template-1">
    <div id="template-host-1">
        <template shadowrootmode="open" shadowrootclonable><x-foo><x-foo></x-foo></x-foo></template>
    </div>
</template>

<template id="template-2">
    <div id="template-host-2">
        <template shadowrootmode="open" shadowrootclonable shadowrootcustomelementregistry><x-foo><x-foo></x-foo></x-foo></template>
    </div>
</template>

<script>
class XFoo extends HTMLElement {}

test(() => {
    assert_equals((new Document).createElement('a-b').customElementRegistry, null);
    assert_equals(document.implementation.createHTMLDocument().createElement('a-b').customElementRegistry, null);
    assert_equals(document.implementation.createDocument('http://www.w3.org/1999/xhtml', 'html', null).createElement('a-b').customElementRegistry, null);
}, 'customElementRegistry of an upgrade candidate created in a document without a browsing context uses null regsitry by default');

function makeIframe(test) {
    const iframe = document.createElement('iframe');
    document.body.appendChild(iframe);
    test.add_cleanup(() => iframe.remove());
    return [iframe.contentDocument, iframe.contentWindow];
}

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>');
    const hostClone = doc.getElementById('host').cloneNode(true);
    assert_equals(hostClone.shadowRoot.customElementRegistry, null);
    assert_equals(hostClone.shadowRoot.querySelector('a-b').customElementRegistry, null);
}, 'Connecting a custom element candiate in a shadow root with a scoped custom element registry has null registry by default');

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>');
    win.customElements.define('a-b', class ABElement extends win.HTMLElement { });
    const hostClone = doc.getElementById('host').cloneNode(true);
    assert_equals(hostClone.shadowRoot.customElementRegistry, null);
    const candidate = hostClone.shadowRoot.querySelector('a-b');
    assert_equals(candidate.customElementRegistry, null);
    doc.body.appendChild(candidate);
    assert_equals(candidate.customElementRegistry, null);
}, 'Connecting a custom element candiate with null registry does not set the registry');

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>');
    win.customElements.define('a-b', class ABElement extends win.HTMLElement { });
    const hostClone = doc.getElementById('host').cloneNode(true);
    const candidate = hostClone.shadowRoot.querySelector('a-b');
    const registry = new CustomElementRegistry;
    assert_equals(candidate.customElementRegistry, null);
    registry.initialize(hostClone.shadowRoot);
    assert_equals(candidate.customElementRegistry, registry);
    doc.body.appendChild(candidate);
    assert_equals(candidate.customElementRegistry, registry);

    const element = doc.createElement('host', {customElementRegistry: registry});
    assert_equals(element.customElementRegistry, registry);
    doc.body.appendChild(element);
    assert_equals(element.customElementRegistry, registry);
}, 'Connecting a custom element candiate with a scoped custom element registry does not change the registry');

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host1"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>'
        + '<div id="host2"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><c-d></c-d></template></div>');
    win.customElements.define('a-b', class ABElement extends win.HTMLElement { });
    const clone1 = doc.getElementById('host1').cloneNode(true);
    const clone2 = doc.getElementById('host2').cloneNode(true);
    const candidate = clone1.shadowRoot.querySelector('a-b');
    assert_equals(candidate.customElementRegistry, null);
    assert_equals(clone2.shadowRoot.querySelector('c-d').customElementRegistry, null);
    const registry = new CustomElementRegistry;
    registry.initialize(clone2.shadowRoot);
    assert_equals(clone2.shadowRoot.querySelector('c-d').customElementRegistry, registry);
    clone2.shadowRoot.appendChild(candidate);
    assert_equals(candidate.customElementRegistry, null);
}, 'Inserting a custom element candiate with null registry does not change the registry');

test((test) => {
    const [doc, win] = makeIframe(test);
    const registry = new CustomElementRegistry;
    const host = doc.createElement('div');
    const shadowRoot = host.attachShadow({mode: 'closed', customElementRegistry: registry});
    assert_equals(shadowRoot.customElementRegistry, registry);
    doc.body.appendChild(host);
    assert_equals(shadowRoot.customElementRegistry, registry);
}, 'Inserting the shadow host of a shadow root with a scoped custom element registry does not change the registry');

test((test) => {
    const [doc1, win1] = makeIframe(test);
    const [doc2, win2] = makeIframe(test);
    class ABElement extends win2.HTMLElement { };
    win2.customElements.define('a-b', ABElement);
    const elementFromDoc1 = doc1.createElement('a-b');
    assert_equals(elementFromDoc1.customElementRegistry, win1.customElements);
    doc2.body.appendChild(elementFromDoc1);
    assert_equals(elementFromDoc1.customElementRegistry, win2.customElements);
    assert_true(elementFromDoc1 instanceof ABElement);
}, 'Inserting a node from another document with global registry results in the custom element registry to be set and upgraded.');

test((test) => {
    const contentDocument = document.implementation.createHTMLDocument();
    class ABElement extends HTMLElement { }
    customElements.define('a-b', ABElement);
    const elementFromContentDoc = contentDocument.createElement('a-b');
    assert_equals(elementFromContentDoc.customElementRegistry, null);
    document.body.appendChild(elementFromContentDoc);
    assert_equals(elementFromContentDoc.customElementRegistry, customElements);
    assert_true(elementFromContentDoc instanceof ABElement);
}, 'Inserting a node from another document with null registry results in the custom element registry to be set and upgraded.');

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>'
        + '<template id="t"><a-b></a-b></template>');
    const registry = new win.CustomElementRegistry();
    const hostClone = doc.getElementById('host').cloneNode(true);
    registry.initialize(hostClone.shadowRoot);
    doc.body.appendChild(hostClone);

    // Clone from template (null registry, no browsing context).
    const templateContent = doc.getElementById('t').content;
    const cloned = templateContent.firstElementChild.cloneNode(true);
    assert_equals(cloned.customElementRegistry, null,
        'Cloned element from template should have null registry');

    // Append into scoped shadow root.
    hostClone.shadowRoot.appendChild(cloned);
    assert_equals(cloned.customElementRegistry, win.customElements,
        'Element from template appended into scoped shadow root should get global registry');
}, 'Inserting a node cloned from a template with null registry into a scoped shadow root should set global registry');

test((test) => {
    const [doc, win] = makeIframe(test);
    doc.documentElement.setHTMLUnsafe('<div id="host"><template shadowrootmode="open" shadowrootclonable="true" shadowrootcustomelementregistry><a-b></a-b></template></div>'
        + '<template id="t"><a-b></a-b></template>');
    const registry = new win.CustomElementRegistry();
    const hostClone = doc.getElementById('host').cloneNode(true);
    registry.initialize(hostClone.shadowRoot);
    doc.body.appendChild(hostClone);

    // Clone from template (null registry), adopt into main doc first.
    const templateContent = doc.getElementById('t').content;
    const cloned = templateContent.firstElementChild.cloneNode(true);
    assert_equals(cloned.customElementRegistry, null,
        'Cloned element from template should have null registry');
    doc.adoptNode(cloned);
    assert_equals(cloned.customElementRegistry, win.customElements,
        'After adoptNode, element should have global registry');

    // Append into scoped shadow root.
    hostClone.shadowRoot.appendChild(cloned);
    assert_equals(cloned.customElementRegistry, win.customElements,
        'Element adopted then appended into scoped shadow root should keep global registry');
}, 'adoptNode then inserting into a scoped shadow root should preserve global registry');

test(t => {
    const template = document.getElementById('template-1');
    const clone = template.content.cloneNode(true);
    const [doc, win] = makeIframe(t);
    doc.body.appendChild(clone);
    const host1shadow = doc.querySelector('#template-host-1').shadowRoot;
    assert_equals(host1shadow.customElementRegistry, win.customElements);
}, "Declarative shadow DOM without shadowrootcustomelementregistry attribute without registry initialized should gain effective global registry after adoption.")

test(t => {
    const template = document.getElementById('template-2');
    const clone = template.content.cloneNode(true);
    const [doc, win] = makeIframe(t);
    doc.body.append(clone);
    const host2shadow = doc.querySelector('#template-host-2').shadowRoot;
    assert_equals(host2shadow.customElementRegistry, null);
}, "Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.")

test(t => {
    const template = document.getElementById('template-1');
    const clone = template.content.cloneNode(true);
    const [doc, win] = makeIframe(t);
    win.customElements.define('x-foo', XFoo);
    doc.body.append(clone);
    const host1shadow = doc.querySelector('#template-host-1').shadowRoot;
    assert_equals(host1shadow.querySelector('x-foo').customElementRegistry, win.customElements);
    assert_true(host1shadow.querySelector('x-foo') instanceof XFoo);
    assert_equals(host1shadow.querySelector('x-foo').querySelector('x-foo').customElementRegistry, win.customElements);
    assert_true(host1shadow.querySelector('x-foo').querySelector('x-foo') instanceof XFoo);
}, "Null registry element should gain effective global registry of global registry from its parent upon adoption.")

test(t => {
    const template = document.getElementById('template-2');
    const clone = template.content.cloneNode(true);
    const [doc, win] = makeIframe(t);
    win.customElements.define('x-foo', XFoo);
    doc.body.append(clone);
    const host2shadow = doc.querySelector('#template-host-2').shadowRoot;
    assert_equals(host2shadow.querySelector('x-foo').customElementRegistry, null);
    assert_false(host2shadow.querySelector('x-foo') instanceof XFoo);
    assert_equals(host2shadow.querySelector('x-foo').querySelector('x-foo').customElementRegistry, null);
    assert_false(host2shadow.querySelector('x-foo').querySelector('x-foo') instanceof XFoo);
}, "Null registry element should gain effective global registry of null from its parent upon adoption.")

test(t => {
    const template = document.getElementById('template-2');
    const clone = template.content.cloneNode(true);
    const registry = new CustomElementRegistry;
    registry.initialize(clone.querySelector('#template-host-2').shadowRoot);
    clone.querySelector('#template-host-2').shadowRoot.appendChild(document.createElement('new-foo', {customElementRegistry: null}));
    clone.querySelector('#template-host-2').shadowRoot.querySelector('x-foo').appendChild(document.createElement('new-foo', {customElementRegistry: null}));
    const [doc, win] = makeIframe(t);
    doc.body.append(clone);
    const host2shadow = doc.querySelector('#template-host-2').shadowRoot;
    assert_equals(host2shadow.querySelector('new-foo').customElementRegistry, null);
    assert_equals(host2shadow.querySelector('x-foo').querySelector('new-foo').customElementRegistry, null);
}, "Null registry element should gain effective global registry of scoped registry from its parent upon adoption.")

test(t => {
    const template = document.getElementById('template-2');
    const clone = template.content.cloneNode(true);
    const registry = new CustomElementRegistry;
    registry.initialize(clone.querySelector('#template-host-2').shadowRoot);
    clone.querySelector('#template-host-2').shadowRoot.appendChild(document.createElement('new-foo'));
    clone.querySelector('#template-host-2').shadowRoot.querySelector('x-foo').appendChild(document.createElement('new-foo'));
    const [doc, win] = makeIframe(t);
    doc.body.append(clone);
    const host2shadow = doc.querySelector('#template-host-2').shadowRoot;
    assert_equals(host2shadow.querySelector('new-foo').customElementRegistry, win.customElements);
    assert_equals(host2shadow.querySelector('x-foo').querySelector('new-foo').customElementRegistry, win.customElements);
}, "Global registry element should gain effective global registry of scoped registry from its parent upon adoption.")
</script>
</body>
</html>
